+++ hermes/bin/src/runtime_extensions/wasi/random/insecure/mod.rs
@@ -1,4 +1,4 @@
-//! CLI runtime extension implementation.
+//! Insecure RNG runtime extension implementation.
 
 use crate::runtime_extensions::state::{Context, Stateful};
