+++ hermes/bin/src/runtime_extensions/wasi/random/insecure_seed/mod.rs
@@ -1,4 +1,4 @@
-//! CLI runtime extension implementation.
+//! Insecure RNG seed runtime extension implementation.
 
 use crate::runtime_extensions::state::{Context, Stateful};
